requirements.txt
asgiref==3.2.3
certifi==2019.11.28
Django==3.0.2
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.5
django-minio-backend==1.1.2
django-minio-storage==0.3.5
django-nine==0.2.3
djangorestframework==3.11.0
elasticsearch==7.5.1
elasticsearch-dsl==7.1.0
Markdown==3.1.1
minio==5.0.6
Pillow==7.0.0
psycopg2==2.8.4
Pygments==2.5.2
python-dateutil==2.8.1
python-decouple==3.3
pytz==2019.3
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
